+++ xknx/remote_value/remote_value_color_rgb.py
@@ -6,11 +6,10 @@ DPT 232.600.
 
 from __future__ import annotations
 
-from collections.abc import Sequence
 from typing import TYPE_CHECKING
 
 from xknx.dpt import DPTArray, DPTBinary
-from xknx.exceptions import ConversionError, CouldNotParseTelegram
+from xknx.dpt.dpt_color import DPTColorRGB, RGBColor
 
 from .remote_value import AsyncCallbackType, GroupAddressesType, RemoteValue
 
@@ -18,7 +17,7 @@ if TYPE_CHECKING:
     from xknx.xknx import XKNX
 
 
-class RemoteValueColorRGB(RemoteValue[tuple[int, int, int]]):
+class RemoteValueColorRGB(RemoteValue[RGBColor]):
     """Abstraction for remote value of KNX DPT 232.600 (DPT_Color_RGB)."""
 
     def __init__(
@@ -42,34 +41,10 @@ class RemoteValueColorRGB(RemoteValue[tuple[int, int, int]]):
             after_update_cb=after_update_cb,
         )
 
-    def to_knx(self, value: Sequence[int]) -> DPTArray:
+    def to_knx(self, value: RGBColor) -> DPTArray:
         """Convert value to payload."""
-        if not isinstance(value, (list, tuple)):
-            raise ConversionError(
-                "Could not serialize RemoteValueColorRGB (wrong type)",
-                value=value,
-                type=type(value),
-            )
-        if len(value) != 3:
-            raise ConversionError(
-                "Could not serialize DPT 232.600 (wrong length)",
-                value=value,
-                type=type(value),
-            )
-        if (
-            any(not isinstance(color, int) for color in value)
-            or any(color < 0 for color in value)
-            or any(color > 255 for color in value)
-        ):
-            raise ConversionError(
-                "Could not serialize DPT 232.600 (wrong bytes)", value=value
-            )
+        return DPTColorRGB.to_knx(value)
 
-        return DPTArray(list(value))
-
-    def from_knx(self, payload: DPTArray | DPTBinary) -> tuple[int, int, int]:
+    def from_knx(self, payload: DPTArray | DPTBinary) -> RGBColor:
         """Convert current payload to value."""
-        if not (isinstance(payload, DPTArray) and len(payload.value) == 3):
-            raise CouldNotParseTelegram("Payload invalid", payload=str(payload))
-
-        return payload.value[0], payload.value[1], payload.value[2]
+        return DPTColorRGB.from_knx(payload)